refactor(app): refactored multiplayer flow
