+++ app/src/components/game/Footer.tsx
@@ -1,10 +1,8 @@
 'use client';
-import { GamesContext } from '@/src/context/GamesContext';
 import { useBlockNumber } from '@/src/hooks/useBlockNumber';
 import { hasGameStarted, isGameFinished } from '@/src/utils/game';
-import { useContext, useEffect, useMemo, useState } from 'react';
-import { OnChainGame } from '../../types/Game';
-import { usePrivyWalletAddress } from '../../hooks/usePrivyWalletAddress';
+import { useGameAndPuzzleData } from '../../hooks/useGameAndPuzzleData';
+import { useCurrentRound } from '../../hooks/useCurrentRound';
 
 export function Footer({
   gameId,
@@ -16,19 +14,19 @@ export function Footer({
   opponentScore: number;
 }) {
   const blockNumber = useBlockNumber();
-  const { games } = useContext(GamesContext);
-  const address = usePrivyWalletAddress();
-  const game = useMemo(() => {
-    return games.find((game) => game.id === gameId);
-  }, [games, gameId]);
+  const {
+    data: { onChainGame },
+  } = useGameAndPuzzleData(gameId);
+  const currentRound = useCurrentRound(onChainGame);
 
-  if (!game || !blockNumber) return null;
-  if (isGameFinished(blockNumber, game)) return null;
-  if (!hasGameStarted(blockNumber, game)) return null;
+  if (!onChainGame || !blockNumber || !currentRound) return null;
+  if (isGameFinished(blockNumber, onChainGame)) return null;
+  if (!hasGameStarted(blockNumber, onChainGame)) return null;
 
   const blocksLeftThisTurn =
-    game.interval -
-    ((Number(blockNumber) - Number(game.startingBlock)) % game.interval);
+    onChainGame.interval -
+    ((Number(blockNumber) - Number(onChainGame.startingBlock)) %
+      onChainGame.interval);
 
   return (
     <footer className="bg-black h-20 mt-auto">
@@ -59,13 +57,7 @@ export function Footer({
             round
             <br />
             <span className="text-xl font-normal">
-              {Math.floor(
-                (Number(blockNumber) - Number(game.startingBlock)) /
-                  game.interval
-              ) +
-                1 +
-                ' of ' +
-                game.numberOfTurns}
+              {currentRound + ' of ' + onChainGame.numberOfRounds}
             </span>
           </h4>
         </div>
